requirements.txt
# requirements.txt
# Telegram Bot Framework
python-telegram-bot==20.7

# LLM Integration
openai==1.35.7
anthropic==0.25.8
langchain==0.2.5
langchain-community==0.2.5
tiktoken

# WooCommerce Integration
woocommerce==3.0.0
woocommerceaio==1.0.0
requests==2.31.0

# Excel Integration
openpyxl==3.1.2
xlsxwriter==3.1.9
pandas==2.1.4
msgraph-sdk
azure-identity==1.16.1

# Data Validation and Models
pydantic==2.7.1
pydantic-settings==2.2.1
marshmallow==3.21.1

# Async Support
asyncio==3.4.3
aiohttp==3.9.5
aiofiles==23.2.0

# Database and Caching
redis==5.0.4
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite

# Logging and Monitoring
structlog==24.1.0
colorlog==6.8.2
python-json-logger==2.0.7
prometheus-client==0.20.0

# Configuration Management
python-dotenv==1.0.1
pyyaml==6.0.1
toml==0.10.2

# File and Media Processing
pillow==10.3.0
moviepy==1.0.3
python-magic==0.4.27
ffmpeg-python==0.2.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1
cryptography==42.0.7
jsonschema==4.22.0
retry==0.9.2

# Testing
pytest==8.2.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-mock==3.14.0
factory-boy==3.3.0

# Development Tools
black==24.4.2
flake8==7.0.0
mypy==1.10.0
isort==5.13.2
pre-commit==3.7.1
